remove class wrapper for client interface
